split and rename BackendHelper to QuoteHelper and StorageS3Helper
