+++ src/Backend/Teradata/Helper/QuoteHelper.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\Db\ImportExport\Backend\Teradata\Helper;
+
+final class QuoteHelper
+{
+    public static function quoteValue(string $value): string
+    {
+        return "'" . addslashes($value) . "'";
+    }
+}
