+++ src/Backend/Teradata/ToStage/FromS3TPTAdapter.php
@@ -6,7 +6,7 @@ namespace Keboola\Db\ImportExport\Backend\Teradata\ToStage;
 
 use Doctrine\DBAL\Connection;
 use Keboola\Db\ImportExport\Backend\CopyAdapterInterface;
-use Keboola\Db\ImportExport\Backend\Teradata\Helper\BackendHelper;
+use Keboola\Db\ImportExport\Backend\Teradata\Helper\StorageS3Helper;
 use Keboola\Db\ImportExport\Backend\Teradata\TeradataImportOptions;
 use Keboola\Db\ImportExport\Backend\Teradata\ToFinalTable\SqlBuilder;
 use Keboola\Db\ImportExport\Backend\Teradata\ToStage\Exception\FailedTPTLoadException;
@@ -190,14 +190,14 @@ class FromS3TPTAdapter implements CopyAdapterInterface
         // b/ it is list of real CSV files (even a single file)
         // c/ it is list of Fxx files generated by TD
 
-        if (BackendHelper::isMultipartFile($source)) {
+        if (StorageS3Helper::isMultipartFile($source)) {
             // scenario c
 
             // file is s3://bucket/path/.../file.csv.gz/F000000
             // extract real filepath from path -> remove F000000 and do just one load
             // load with SPF = false
 
-            [$prefix, $object] = BackendHelper::buildPrefixAndObject($source);
+            [$prefix, $object] = StorageS3Helper::buildPrefixAndObject($source);
             $moduleStr = sprintf(
             // phpcs:ignore
                 'AccessModuleInitStr = \'S3Region="%s" S3Bucket="%s" S3Prefix="%s" S3Object="%s" S3SinglePartFile=False S3ConfigDir=%s\'',
@@ -212,7 +212,7 @@ class FromS3TPTAdapter implements CopyAdapterInterface
                 // load with wildcard
                 // scenario b
 
-                $mask = BackendHelper::getMask($source);
+                $mask = StorageS3Helper::getMask($source);
                 $path = RelativePath::createFromRootAndPath(new S3Provider(), $source->getBucket(), $mask);
                 $moduleStr = sprintf(
                 // phpcs:ignore
@@ -282,7 +282,7 @@ EOD;
         $ignoredLines = $importOptions->getNumberOfIgnoredLines();
 
         $quotedDestination = TeradataQuote::quoteSingleIdentifier($destination->getSchemaName());
-        $tablesPrefix = BackendHelper::generateStagingTableName();
+        $tablesPrefix = StorageS3Helper::generateStagingTableName();
         $logTable = $tablesPrefix . '_log';
         $logTableQuoted = $quotedDestination . '.' . TeradataQuote::quoteSingleIdentifier($logTable);
         $errTable1 = $tablesPrefix . '_e1';
